chore: move /pudgy to /cold-storage-sbt
