+++ app/[locale]/cold-storage-sbt/PudgyChecker.tsx
