+++ app/[locale]/cold-storage-sbt/PudgyQuiz.tsx
