+++ app/[locale]/cold-storage-sbt/PudgyShareButton.tsx
