+++ app/[locale]/cold-storage-sbt/page.tsx
